+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2024-03-14
+
 ### Added
 
 * The RPC interface now offers a facility for sending custom user errors through `send_custom_error` and `send_custom_error_and_shutdown`.